+++ src/dags/custom_node.cpp
@@ -48,7 +48,7 @@ CustomNode::CustomNode(
 
 Status CustomNode::execute(session_key_t sessionKey, PipelineEventQueue& notifyEndQueue) {
     auto& nodeSession = getNodeSession(sessionKey);
-    auto& customNodeSession = static_cast<CustomNodeSession&>(nodeSession);
+    CustomNodeSession& customNodeSession = static_cast<CustomNodeSession&>(nodeSession);
     return customNodeSession.execute(notifyEndQueue, *this, this->library, this->libraryParameters, this->parameters.size(), getCNLIMWrapperPtr(customNodeLibraryInternalManager));
 }
