+++ src/dags/gathernodeinputhandler.hpp
@@ -26,6 +26,7 @@
 #include "../shape.hpp"
 #include "nodeinputhandler.hpp"
 #include "session_id.hpp"
+#include "nodesessionmetadata.hpp"
 
 namespace ovms {
 
@@ -38,7 +39,16 @@ class GatherNodeInputHandler : public NodeInputHandler {
     std::unique_ptr<CollapseDetails> collapsingDetails;
 
 public:
-    GatherNodeInputHandler(uint32_t inputsMissingCount, const CollapseDetails& collapsingDetails);
+    // TODO: Investigate why windows does not see this symbol
+    GatherNodeInputHandler(uint32_t inputsMissingCount, const CollapseDetails& collapsingDetails) :
+    NodeInputHandler(inputsMissingCount),
+    collapsingDetails(std::make_unique<CollapseDetails>(collapsingDetails)) {
+    remainingDependencies = std::accumulate(
+        collapsingDetails.collapsedSessionSizes.begin(),
+        collapsingDetails.collapsedSessionSizes.end(),
+        remainingDependencies,
+        std::multiplies<session_id_t>());
+    }
     Status setInput(const std::string& inputName, TensorWithSource& tensor, session_id_t shardId) override;
     Status notifyFinishedDependency() override;
