+++ src/tensor_conversion.cpp
@@ -101,13 +101,13 @@ static Status convertPrecision(const cv::Mat& src, cv::Mat& dst, const ovms::Pre
     return StatusCode::OK;
 }
 
-static Status validateLayout(const std::shared_ptr<const TensorInfo>& tensorInfo) {
+static Status validateLayout(const TensorInfo& tensorInfo) {
     OVMS_PROFILE_FUNCTION();
     static const std::string binarySupportedLayout = "N...HWC";
-    if (!tensorInfo->getLayout().createIntersection(Layout(binarySupportedLayout), tensorInfo->getShape().size()).has_value()) {
+    if (!tensorInfo.getLayout().createIntersection(Layout(binarySupportedLayout), tensorInfo.getShape().size()).has_value()) {
         SPDLOG_DEBUG("Endpoint needs to be compatible with {} to support binary image inputs, actual: {}",
             binarySupportedLayout,
-            tensorInfo->getLayout());
+            tensorInfo.getLayout());
         return StatusCode::UNSUPPORTED_LAYOUT;
     }
     return StatusCode::OK;
@@ -126,17 +126,17 @@ static Status resizeMat(const cv::Mat& src, cv::Mat& dst, const dimension_value_
     return StatusCode::OK;
 }
 
-static Status validateNumberOfChannels(const std::shared_ptr<const TensorInfo>& tensorInfo,
+static Status validateNumberOfChannels(const TensorInfo& tensorInfo,
     const cv::Mat input,
     cv::Mat* firstBatchImage) {
     OVMS_PROFILE_FUNCTION();
 
     // At this point we can either have nhwc format or pretendant to be nhwc but with ANY layout in pipeline info
     Dimension numberOfChannels;
-    if (tensorInfo->getShape().size() == 4) {
-        numberOfChannels = tensorInfo->getShape()[3];
-    } else if (tensorInfo->isInfluencedByDemultiplexer() && tensorInfo->getShape().size() == 5) {
-        numberOfChannels = tensorInfo->getShape()[4];
+    if (tensorInfo.getShape().size() == 4) {
+        numberOfChannels = tensorInfo.getShape()[3];
+    } else if (tensorInfo.isInfluencedByDemultiplexer() && tensorInfo.getShape().size() == 5) {
+        numberOfChannels = tensorInfo.getShape()[4];
     } else {
         return StatusCode::INVALID_NO_OF_CHANNELS;
     }
@@ -148,7 +148,7 @@ static Status validateNumberOfChannels(const std::shared_ptr<const TensorInfo>& 
     }
     if (!numberOfChannels.match(input.channels())) {
         SPDLOG_DEBUG("Binary data sent to input: {} has invalid number of channels. Expected: {} Actual: {}",
-            tensorInfo->getMappedName(),
+            tensorInfo.getMappedName(),
             numberOfChannels.toString(),
             input.channels());
         return StatusCode::INVALID_NO_OF_CHANNELS;
@@ -167,16 +167,16 @@ static Status validateResolutionAgainstFirstBatchImage(const cv::Mat input, cv::
     return StatusCode::BINARY_IMAGES_RESOLUTION_MISMATCH;
 }
 
-static bool checkBatchSizeMismatch(const std::shared_ptr<const TensorInfo>& tensorInfo,
+static bool checkBatchSizeMismatch(const TensorInfo& tensorInfo,
     const int batchSize) {
     OVMS_PROFILE_FUNCTION();
-    if (!tensorInfo->getBatchSize().has_value() || batchSize == 0) {
+    if (!tensorInfo.getBatchSize().has_value() || batchSize == 0) {
         return true;
     }
-    return !tensorInfo->getBatchSize().value().match(batchSize);
+    return !tensorInfo.getBatchSize().value().match(batchSize);
 }
 
-static Status validateInput(const std::shared_ptr<const TensorInfo>& tensorInfo, const cv::Mat input, cv::Mat* firstBatchImage, bool enforceResolutionAlignment) {
+static Status validateInput(const TensorInfo& tensorInfo, const cv::Mat input, cv::Mat* firstBatchImage, bool enforceResolutionAlignment) {
     // Binary inputs are supported for any endpoint that is compatible with N...HWC layout.
     // With unknown layout, there is no way to deduce expected endpoint input resolution.
     // This forces binary utility to create tensors with resolution inherited from first batch of binary input image (request).
@@ -192,7 +192,7 @@ static Status validateInput(const std::shared_ptr<const TensorInfo>& tensorInfo,
     return validateNumberOfChannels(tensorInfo, input, firstBatchImage);
 }
 
-static Status validateTensor(const std::shared_ptr<const TensorInfo>& tensorInfo,
+static Status validateTensor(const TensorInfo& tensorInfo,
     const tensorflow::TensorProto& src,
     const std::string* buffer) {
     OVMS_PROFILE_FUNCTION();
@@ -201,16 +201,16 @@ static Status validateTensor(const std::shared_ptr<const TensorInfo>& tensorInfo
         return status;
     }
     // 4 for default pipelines, 5 for pipelines with demultiplication at entry
-    bool isShapeLengthValid = tensorInfo->getShape().size() == 4 ||
-                              (tensorInfo->isInfluencedByDemultiplexer() && tensorInfo->getShape().size() == 5);
+    bool isShapeLengthValid = tensorInfo.getShape().size() == 4 ||
+                              (tensorInfo.isInfluencedByDemultiplexer() && tensorInfo.getShape().size() == 5);
     if (!isShapeLengthValid) {
         return StatusCode::INVALID_SHAPE;
     }
 
     if (checkBatchSizeMismatch(tensorInfo, src.string_val_size())) {
         SPDLOG_DEBUG("Input: {} request batch size is incorrect. Expected: {} Actual: {}",
-            tensorInfo->getMappedName(),
-            tensorInfo->getBatchSize().has_value() ? tensorInfo->getBatchSize().value().toString() : std::string{"none"},
+            tensorInfo.getMappedName(),
+            tensorInfo.getBatchSize().has_value() ? tensorInfo.getBatchSize().value().toString() : std::string{"none"},
             src.string_val_size());
         return StatusCode::INVALID_BATCH_SIZE;
     }
@@ -232,7 +232,7 @@ inline static int getNumberOfInputs(const std::string* buffer) {
     return batchSize;
 }
 
-static Status validateTensor(const std::shared_ptr<const TensorInfo>& tensorInfo,
+static Status validateTensor(const TensorInfo& tensorInfo,
     const ::KFSRequest::InferInputTensor& src,
     const std::string* buffer) {
     OVMS_PROFILE_FUNCTION();
@@ -242,8 +242,8 @@ static Status validateTensor(const std::shared_ptr<const TensorInfo>& tensorInfo
         return status;
     }
     // 4 for default pipelines, 5 for pipelines with demultiplication at entry
-    bool isShapeLengthValid = tensorInfo->getShape().size() == 4 ||
-                              (tensorInfo->isInfluencedByDemultiplexer() && tensorInfo->getShape().size() == 5);
+    bool isShapeLengthValid = tensorInfo.getShape().size() == 4 ||
+                              (tensorInfo.isInfluencedByDemultiplexer() && tensorInfo.getShape().size() == 5);
     if (!isShapeLengthValid) {
         return StatusCode::INVALID_SHAPE;
     }
@@ -251,8 +251,8 @@ static Status validateTensor(const std::shared_ptr<const TensorInfo>& tensorInfo
     size_t batchSize = !rawInputsContentsUsed ? src.contents().bytes_contents_size() : getNumberOfInputs(buffer);
     if (checkBatchSizeMismatch(tensorInfo, batchSize)) {
         SPDLOG_DEBUG("Input: {} request batch size is incorrect. Expected: {} Actual: {}",
-            tensorInfo->getMappedName(),
-            tensorInfo->getBatchSize().has_value() ? tensorInfo->getBatchSize().value().toString() : std::string{"none"},
+            tensorInfo.getMappedName(),
+            tensorInfo.getBatchSize().has_value() ? tensorInfo.getBatchSize().value().toString() : std::string{"none"},
             src.contents().bytes_contents_size());
         return StatusCode::INVALID_BATCH_SIZE;
     }
@@ -274,22 +274,22 @@ static Status validateTensor(const std::shared_ptr<const TensorInfo>& tensorInfo
     return StatusCode::OK;
 }
 
-static Dimension getTensorInfoHeightDim(const std::shared_ptr<const TensorInfo>& tensorInfo) {
-    size_t numberOfShapeDimensions = tensorInfo->getShape().size();
+static Dimension getTensorInfoHeightDim(const TensorInfo& tensorInfo) {
+    size_t numberOfShapeDimensions = tensorInfo.getShape().size();
     if (numberOfShapeDimensions < 4 || numberOfShapeDimensions > 5) {
         throw std::logic_error("wrong number of shape dimensions");
     }
     size_t position = numberOfShapeDimensions == 4 ? /*NHWC*/ 1 : /*N?HWC*/ 2;
-    return tensorInfo->getShape()[position];
+    return tensorInfo.getShape()[position];
 }
 
-static Dimension getTensorInfoWidthDim(const std::shared_ptr<const TensorInfo>& tensorInfo) {
-    size_t numberOfShapeDimensions = tensorInfo->getShape().size();
+static Dimension getTensorInfoWidthDim(const TensorInfo& tensorInfo) {
+    size_t numberOfShapeDimensions = tensorInfo.getShape().size();
     if (numberOfShapeDimensions < 4 || numberOfShapeDimensions > 5) {
         throw std::logic_error("wrong number of shape dimensions");
     }
     size_t position = numberOfShapeDimensions == 4 ? /*NHWC*/ 2 : /*N?HWC*/ 3;
-    return tensorInfo->getShape()[position];
+    return tensorInfo.getShape()[position];
 }
 
 static void updateTargetResolution(Dimension& height, Dimension& width, const cv::Mat& image) {
@@ -321,15 +321,15 @@ static void updateTargetResolution(Dimension& height, Dimension& width, const cv
     }
 }
 
-static bool isResizeSupported(const std::shared_ptr<const TensorInfo>& tensorInfo) {
-    for (const auto& dim : tensorInfo->getShape()) {
+static bool isResizeSupported(const TensorInfo& tensorInfo) {
+    for (const auto& dim : tensorInfo.getShape()) {
         if (dim.isAny()) {
             return false;
         }
     }
-    if (tensorInfo->getLayout() != "NHWC" &&
-        tensorInfo->getLayout() != "N?HWC" &&
-        tensorInfo->getLayout() != Layout::getUnspecifiedLayout()) {
+    if (tensorInfo.getLayout() != "NHWC" &&
+        tensorInfo.getLayout() != "N?HWC" &&
+        tensorInfo.getLayout() != Layout::getUnspecifiedLayout()) {
         return false;
     }
     return true;
@@ -371,7 +371,7 @@ inline static Status getInputs(const std::string* buffer, std::vector<std::strin
 }
 
 template <typename TensorType>
-static Status convertTensorToMatsMatchingTensorInfo(const TensorType& src, std::vector<cv::Mat>& images, const std::shared_ptr<const TensorInfo>& tensorInfo, const std::string* buffer) {
+static Status convertTensorToMatsMatchingTensorInfo(const TensorType& src, std::vector<cv::Mat>& images, const TensorInfo& tensorInfo, const std::string* buffer) {
     OVMS_PROFILE_FUNCTION();
     Dimension targetHeight = getTensorInfoHeightDim(tensorInfo);
     Dimension targetWidth = getTensorInfoWidthDim(tensorInfo);
@@ -400,9 +400,9 @@ static Status convertTensorToMatsMatchingTensorInfo(const TensorType& src, std::
             updateTargetResolution(targetHeight, targetWidth, image);
         }
 
-        if (!isPrecisionEqual(image.depth(), tensorInfo->getPrecision())) {
+        if (!isPrecisionEqual(image.depth(), tensorInfo.getPrecision())) {
             cv::Mat imageCorrectPrecision;
-            status = convertPrecision(image, imageCorrectPrecision, tensorInfo->getPrecision());
+            status = convertPrecision(image, imageCorrectPrecision, tensorInfo.getPrecision());
 
             if (status != StatusCode::OK) {
                 return status;
@@ -432,11 +432,11 @@ static Status convertTensorToMatsMatchingTensorInfo(const TensorType& src, std::
     return StatusCode::OK;
 }
 
-static shape_t getShapeFromImages(const std::vector<cv::Mat>& images, const std::shared_ptr<const TensorInfo>& tensorInfo) {
+static shape_t getShapeFromImages(const std::vector<cv::Mat>& images, const TensorInfo& tensorInfo) {
     OVMS_PROFILE_FUNCTION();
     shape_t dims;
     dims.push_back(images.size());
-    if (tensorInfo->isInfluencedByDemultiplexer()) {
+    if (tensorInfo.isInfluencedByDemultiplexer()) {
         dims.push_back(1);
     }
     dims.push_back(images[0].rows);
@@ -445,10 +445,10 @@ static shape_t getShapeFromImages(const std::vector<cv::Mat>& images, const std:
     return dims;
 }
 
-static ov::Tensor createTensorFromMats(const std::vector<cv::Mat>& images, const std::shared_ptr<const TensorInfo>& tensorInfo) {
+static ov::Tensor createTensorFromMats(const std::vector<cv::Mat>& images, const TensorInfo& tensorInfo) {
     OVMS_PROFILE_FUNCTION();
     ov::Shape shape = getShapeFromImages(images, tensorInfo);
-    ov::element::Type precision = tensorInfo->getOvPrecision();
+    ov::element::Type precision = tensorInfo.getOvPrecision();
     ov::Tensor tensor(precision, shape);
     char* ptr = (char*)tensor.data();
     for (cv::Mat image : images) {
@@ -458,9 +458,9 @@ static ov::Tensor createTensorFromMats(const std::vector<cv::Mat>& images, const
     return tensor;
 }
 
-static ov::Tensor convertMatsToTensor(std::vector<cv::Mat>& images, const std::shared_ptr<const TensorInfo>& tensorInfo) {
+static ov::Tensor convertMatsToTensor(std::vector<cv::Mat>& images, const TensorInfo& tensorInfo) {
     OVMS_PROFILE_FUNCTION();
-    switch (tensorInfo->getPrecision()) {
+    switch (tensorInfo.getPrecision()) {
     case ovms::Precision::FP32:
     case ovms::Precision::I32:
     case ovms::Precision::FP64:
@@ -481,7 +481,7 @@ static ov::Tensor convertMatsToTensor(std::vector<cv::Mat>& images, const std::s
 }
 
 template <typename TensorType>
-static Status convertNativeFileFormatRequestTensorToOVTensor(const TensorType& src, ov::Tensor& tensor, const std::shared_ptr<const TensorInfo>& tensorInfo, const std::string* buffer) {
+Status convertNativeFileFormatRequestTensorToOVTensor(const TensorType& src, ov::Tensor& tensor, const TensorInfo& tensorInfo, const std::string* buffer) {
     OVMS_PROFILE_FUNCTION();
     auto status = validateTensor(tensorInfo, src, buffer);
     if (status != StatusCode::OK) {
@@ -502,7 +502,7 @@ static Status convertNativeFileFormatRequestTensorToOVTensor(const TensorType& s
     return StatusCode::OK;
 }
 template <>
-Status convertNativeFileFormatRequestTensorToOVTensor(const ovms::InferenceTensor& src, ov::Tensor& tensor, const std::shared_ptr<const TensorInfo>& tensorInfo, const std::string* buffer) {
+Status convertNativeFileFormatRequestTensorToOVTensor(const ovms::InferenceTensor& src, ov::Tensor& tensor, const TensorInfo& tensorInfo, const std::string* buffer) {
     SPDLOG_ERROR("String conversion is not implemented for C-API");
     return StatusCode::NOT_IMPLEMENTED;
 }
@@ -654,9 +654,9 @@ template Status convertStringRequestToOVTensor<tensorflow::TensorProto>(const te
 template Status convertStringRequestToOVTensor<::KFSRequest::InferInputTensor>(const ::KFSRequest::InferInputTensor& src, ov::Tensor& tensor, const std::string* buffer);
 template Status convertStringRequestToOVTensor<ovms::InferenceTensor>(const ovms::InferenceTensor& src, ov::Tensor& tensor, const std::string* buffer);
 
-template Status convertNativeFileFormatRequestTensorToOVTensor<tensorflow::TensorProto>(const tensorflow::TensorProto& src, ov::Tensor& tensor, const std::shared_ptr<const TensorInfo>& tensorInfo, const std::string* buffer);
-template Status convertNativeFileFormatRequestTensorToOVTensor<::KFSRequest::InferInputTensor>(const ::KFSRequest::InferInputTensor& src, ov::Tensor& tensor, const std::shared_ptr<const TensorInfo>& tensorInfo, const std::string* buffer);
-template Status convertNativeFileFormatRequestTensorToOVTensor<ovms::InferenceTensor>(const ovms::InferenceTensor& src, ov::Tensor& tensor, const std::shared_ptr<const TensorInfo>& tensorInfo, const std::string* buffer);
+template Status convertNativeFileFormatRequestTensorToOVTensor<tensorflow::TensorProto>(const tensorflow::TensorProto& src, ov::Tensor& tensor, const TensorInfo& tensorInfo, const std::string* buffer);
+template Status convertNativeFileFormatRequestTensorToOVTensor<::KFSRequest::InferInputTensor>(const ::KFSRequest::InferInputTensor& src, ov::Tensor& tensor, const TensorInfo& tensorInfo, const std::string* buffer);
+template Status convertNativeFileFormatRequestTensorToOVTensor<ovms::InferenceTensor>(const ovms::InferenceTensor& src, ov::Tensor& tensor, const TensorInfo& tensorInfo, const std::string* buffer);
 
 template Status convertStringRequestToOVTensor2D<tensorflow::TensorProto>(const tensorflow::TensorProto& src, ov::Tensor& tensor, const std::string* buffer);
 template Status convertStringRequestToOVTensor2D<::KFSRequest::InferInputTensor>(const ::KFSRequest::InferInputTensor& src, ov::Tensor& tensor, const std::string* buffer);
